+++ src/components/listview/listview.js
@@ -76,8 +76,8 @@ const LISTVIEW_DEFAULTS = {
   },
   searchTermMinSize: 1,
   initializeContents: false,
-  hasFilter: false,
-  hasSort: false,
+  hasFilters: false,
+  filters: [],
   attributes: null,
   attributesOverride: null
 };
@@ -151,39 +151,31 @@ ListView.prototype = {
     const selectable = this.element.attr('data-selectable');
     const selectOnFocus = this.element.attr('data-select-onfocus');
 
-    if (this.settings.hasFilter || this.settings.hasSort) {
+    if (this.settings.hasFilters && this.settings.filters.length > 0) {
       const filterWrapper = $('<div class="listview-filter-wrapper"></div>');
       const listViewWrapper = self.element.siblings('.listview-search');
+      const filters = this.settings.filters;
       listViewWrapper.append(filterWrapper);
       const searchFieldWrapper = listViewWrapper.find('.searchfield-wrapper');
       let filterCount = 0;
 
-      if (this.settings.hasFilter) {
-        const filterButton = `<button type="button" class="btn-icon listview-filters" title="Filter" class="filter">
+      for (let i = 0; i < filters.length; i++) {
+        const filterName = filters[i];
+        const filterButton = `<button type="button" class="btn-icon listview-filters" title="${filterName}" class="filter">
             <svg class="icon filter-icon" focusable="false" aria-hidden="true" role="presentation">
-                <use href="#icon-filter"></use>
+                <use href="#icon-${filterName}"></use>
             </svg>
-            <span class="audible">Filter</span>
+            <span class="audible">${filterName}</span>
         </button>`;
+
         filterWrapper.append(filterButton);
         filterCount++;
       }
-      
-      if (this.settings.hasSort) {
-        const sortButton = `<button type="button" class="btn-icon listview-filters" title="Filter" class="sort">
-            <svg class="icon filter-icon" focusable="false" aria-hidden="true" role="presentation">
-                <use href="#icon-sort-down"></use>
-            </svg>
-            <span class="audible">Sort</span>
-        </button>`;
-        filterWrapper.append(sortButton);
-        filterCount++;
-      }
 
-      const filterWidth = filterCount > 1 ? 80 : 40;
-      const filterClass = filterCount > 1 ? 'has-listview-filters' : 'has-listview-filter';
+      const filterWidth = filterCount * 40;
       filterWrapper.css('width', `${filterWidth}px`);
-      searchFieldWrapper.addClass(filterClass);
+      searchFieldWrapper.addClass('has-listview-filters');
+      searchFieldWrapper.css('width', `calc(100% - ${filterWidth}px)`);
     }
 
     // Check for legacy data attributes